intent_drift_detector/requirements.txt
# Python dependencies for your agent
# Add any packages your agent needs here

# HTTP client for NDFC API calls
requests>=2.28.0
urllib3>=2.0.0

# YAML parsing for intent files
pyyaml>=6.0

# Example dependencies:
# pandas>=1.5.0
# numpy>=1.21.0
